disable throttle by default
