+++ http-core/src/main/resources/reference.conf
@@ -313,7 +313,8 @@ pekko.http {
         throttle-cost = 100
         throttle-burst = 100
         # setting pekko.http.server.http2.reset-frame.throttle-interval to 0s will disable the throttle
-        throttle-interval = 1s
+        # try changing it to 1s if you want to apply a meaningful limit (the cost and burst apply to the interval time)
+        throttle-interval = 0s
       }
     }
