--- http2-tests/src/test/scala/org/apache/pekko/http/impl/engine/http2/Http2ServerDisableResetSpec.scala
@@ -1,44 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * license agreements; and to You under the Apache License, version 2.0:
- *
- *   https://www.apache.org/licenses/LICENSE-2.0
- *
- * This file is part of the Apache Pekko project, which was derived from Akka.
- */
-
-/*
- * Copyright (C) 2018-2022 Lightbend Inc. <https://www.lightbend.com>
- */
-
-package org.apache.pekko.http.impl.engine.http2
-
-import org.apache.pekko
-import pekko.http.impl.engine.http2.Http2Protocol.FrameType
-import pekko.http.impl.engine.http2.framing.FrameRenderer
-import pekko.util.ByteStringBuilder
-
-import java.nio.ByteOrder
-
-/**
- * This tests the http2 server support for rapid resets can be disabled.
- */
-class Http2ServerDisableResetSpec extends Http2SpecWithMaterializer("""
-    pekko.http.server.remote-address-header = on
-    pekko.http.server.http2.log-frames = on
-    pekko.http.server.http2.reset-frame.throttle-interval = 0s
-  """) {
-  override def failOnSevereMessages: Boolean = true
-
-  "The Http/2 server implementation" should {
-    "not cancel connection during rapid reset attack (throttle disabled)".inAssertAllStagesStopped(
-      new TestSetup with RequestResponseProbes {
-        implicit val bigEndian: ByteOrder = ByteOrder.BIG_ENDIAN
-        val bb = new ByteStringBuilder
-        bb.putInt(0)
-        val rstFrame = FrameRenderer.renderFrame(FrameType.RST_STREAM, ByteFlag.Zero, 1, bb.result())
-        val longFrame = Seq.fill(1000)(rstFrame).reduce(_ ++ _)
-        network.sendBytes(longFrame)
-      })
-  }
-}
