+++ http2-tests/src/test/scala/org/apache/pekko/http/impl/engine/http2/Http2ServerDisableResetThrottleSpec.scala
@@ -21,23 +21,23 @@ import pekko.util.ByteStringBuilder
 import java.nio.ByteOrder
 
 /**
- * This tests the http2 server support for rapid resets.
+ * This tests the http2 server throttle support for rapid resets is disabled by default.
  */
-class Http2ServerResetSpec extends Http2SpecWithMaterializer("""
+class Http2ServerDisableResetThrottleSpec extends Http2SpecWithMaterializer("""
     pekko.http.server.remote-address-header = on
     pekko.http.server.http2.log-frames = on
   """) {
-  override val expectSevereLogsOnlyToMatch: Option[String] = Some(
-    "HTTP2 connection failed with error [Maximum throttle throughput exceeded.]. Sending INTERNAL_ERROR and closing connection.")
+  override def failOnSevereMessages: Boolean = true
 
   "The Http/2 server implementation" should {
-    "cancel connection during rapid reset attack".inAssertAllStagesStopped(new TestSetup with RequestResponseProbes {
-      implicit val bigEndian: ByteOrder = ByteOrder.BIG_ENDIAN
-      val bb = new ByteStringBuilder
-      bb.putInt(0)
-      val rstFrame = FrameRenderer.renderFrame(FrameType.RST_STREAM, ByteFlag.Zero, 1, bb.result())
-      val longFrame = Seq.fill(1000)(rstFrame).reduce(_ ++ _)
-      network.sendBytes(longFrame)
-    })
+    "not cancel connection during rapid reset attack (throttle disabled)".inAssertAllStagesStopped(
+      new TestSetup with RequestResponseProbes {
+        implicit val bigEndian: ByteOrder = ByteOrder.BIG_ENDIAN
+        val bb = new ByteStringBuilder
+        bb.putInt(0)
+        val rstFrame = FrameRenderer.renderFrame(FrameType.RST_STREAM, ByteFlag.Zero, 1, bb.result())
+        val longFrame = Seq.fill(1000)(rstFrame).reduce(_ ++ _)
+        network.sendBytes(longFrame)
+      })
   }
 }
